Move metaworld examples, other refactors
- remove the torch linear feature baseline wrapper
- add param docs to all new functions and launchers
- delete unused docker and gcp launch files

- mt1 examples do not work yet
