+++ metaworld_examples/maml_trpo_metaworld_ml10.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-"""This is an example to train MAML-TRPO on ML45 Push environment."""
+"""This is an example to train MAML-TRPO on ML1 Push environment."""
 # pylint: disable=no-value-for-parameter
 # yapf: disable
 import click
@@ -11,10 +11,10 @@ from garage.envs import MetaWorldSetTaskEnv
 from garage.experiment import (MetaEvaluator, MetaWorldTaskSampler,
                                SetTaskSampler)
 from garage.experiment.deterministic import set_seed
+from garage.np.baselines import LinearFeatureBaseline
 from garage.sampler import RaySampler
 from garage.torch.algos import MAMLTRPO
 from garage.torch.policies import GaussianMLPPolicy
-from garage.torch.value_functions import LinearFeatureValueFunction
 from garage.trainer import Trainer
 
 
@@ -22,11 +22,11 @@ from garage.trainer import Trainer
 @click.command()
 @click.option('--seed', type=int, default=1)
 @click.option('--epochs', type=int, default=2000)
-@click.option('--rollouts_per_task', type=int, default=20)
-@click.option('--meta_batch_size', type=int, default=45)
+@click.option('--rollouts_per_task', type=int, default=10)
+@click.option('--meta_batch_size', type=int, default=20)
 @click.option('--inner_lr', default=1e-4, type=float)
 @wrap_experiment(snapshot_mode='none', name_parameters='passed')
-def maml_trpo_metaworld_ML45(ctxt, seed, epochs, rollouts_per_task,
+def maml_trpo_metaworld_ml10(ctxt, seed, epochs, rollouts_per_task,
                              meta_batch_size, inner_lr):
     """Set up environment and algorithm and run the task.
 
@@ -39,16 +39,22 @@ def maml_trpo_metaworld_ML45(ctxt, seed, epochs, rollouts_per_task,
         rollouts_per_task (int): Number of rollouts per epoch per task
             for training.
         meta_batch_size (int): Number of tasks sampled per batch.
+        inner_lr (float): learning rate to use for the inner TRPO agent.
+            This hyperparameter is typically the one to tune when tuning
+            your MAML.
 
     """
     set_seed(seed)
 
-    ML45 = metaworld.ML45()
-    tasks = MetaWorldTaskSampler(ML45, 'train')
-    env = tasks.sample(45)[0]()
+    ml10 = metaworld.ML10()
+    tasks = MetaWorldTaskSampler(
+        ml10,
+        'train',
+    )
+    env = tasks.sample(10)[0]()
     test_sampler = SetTaskSampler(
         MetaWorldSetTaskEnv,
-        env=MetaWorldSetTaskEnv(ML45, 'test'),
+        env=MetaWorldSetTaskEnv(ml10, 'test'),
     )
     num_test_envs = 5
 
@@ -60,15 +66,17 @@ def maml_trpo_metaworld_ML45(ctxt, seed, epochs, rollouts_per_task,
                                max_std=1.5,
                                std_mlp_type='share_mean_std')
 
-    value_function = LinearFeatureValueFunction(env_spec=env.spec, )
+    value_function = LinearFeatureBaseline(env_spec=env.spec)
 
     meta_evaluator = MetaEvaluator(test_task_sampler=test_sampler,
                                    n_exploration_eps=rollouts_per_task,
                                    n_test_tasks=num_test_envs * 2,
                                    n_test_episodes=10)
+
     sampler = RaySampler(agents=policy,
                          envs=env,
-                         max_episode_length=env.spec.max_episode_length)
+                         max_episode_length=env.spec.max_episode_length,
+                         n_workers=meta_batch_size)
 
     trainer = Trainer(ctxt)
     algo = MAMLTRPO(
@@ -87,7 +95,6 @@ def maml_trpo_metaworld_ML45(ctxt, seed, epochs, rollouts_per_task,
         policy_ent_coeff=5e-5,
         stop_entropy_gradient=True,
         center_adv=False,
-        evaluate_every_n_epochs=10,
     )
 
     trainer.setup(algo, env)
@@ -95,4 +102,4 @@ def maml_trpo_metaworld_ML45(ctxt, seed, epochs, rollouts_per_task,
                   batch_size=rollouts_per_task * env.spec.max_episode_length)
 
 
-maml_trpo_metaworld_ML45()
+maml_trpo_metaworld_ml10()
