+++ metaworld_examples/mtppo_metaworld_mt1.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-"""This is an example to train TRPO on MT10 environment."""
+"""This is an example to train PPO on mt1 environment."""
 # pylint: disable=no-value-for-parameter
 import click
 import metaworld
@@ -7,12 +7,12 @@ import tensorflow as tf
 
 from garage import wrap_experiment
 from garage.envs import normalize
-from garage.envs.multi_env_wrapper import MultiEnvWrapper, round_robin_strategy
 from garage.experiment import MetaWorldTaskSampler
 from garage.experiment.deterministic import set_seed
 from garage.sampler import RaySampler
-from garage.tf.algos import TRPO
+from garage.tf.algos import PPO
 from garage.tf.baselines import GaussianMLPBaseline
+from garage.tf.optimizers import FirstOrderOptimizer
 from garage.tf.policies import GaussianMLPPolicy
 from garage.trainer import TFTrainer
 
@@ -20,8 +20,9 @@ from garage.trainer import TFTrainer
 @click.command()
 @click.option('--seed', default=1)
 @click.option('--entropy', default=5e-3)
+@click.option('--env-name')
 @wrap_experiment(snapshot_mode='none', name_parameters='passed')
-def mttrpo_metaworld_mt10(ctxt, seed, entropy):
+def mtppo_metaworld_mt1(ctxt, seed, entropy, env_name):
     """Set up environment and algorithm and run the task.
 
     Args:
@@ -29,15 +30,16 @@ def mttrpo_metaworld_mt10(ctxt, seed, entropy):
             configuration used by Trainer to create the snapshotter.
         seed (int): Used to seed the random number generator to produce
             determinism.
-        epochs (int): Number of training epochs.
-        batch_size (int): Number of environment steps in one batch.
-        n_tasks (int): Number of tasks to use. Should be a multiple of 10.
+        entropy (float): Coefficient to weigh the entropy reward term by
+            when using the max entropy reward.
+        env_name (str): Name of Meta-World environment to initialize
+            this experiment with.
 
     """
-    n_tasks = 10
+    n_tasks = 50
     set_seed(seed)
-    mt10 = metaworld.MT10()
-    train_task_sampler = MetaWorldTaskSampler(mt10,
+    mt1 = metaworld.MT1(env_name)
+    train_task_sampler = MetaWorldTaskSampler(mt1,
                                               'train',
                                               lambda env, _: normalize(env),
                                               add_env_onehot=True)
@@ -48,7 +50,7 @@ def mttrpo_metaworld_mt10(ctxt, seed, entropy):
     with TFTrainer(snapshot_config=ctxt) as trainer:
         policy = GaussianMLPPolicy(
             env_spec=env.spec,
-            hidden_sizes=(256, 256),
+            hidden_sizes=(512, 512),
             hidden_nonlinearity=tf.nn.tanh,
             output_nonlinearity=None,
             std_share_network=True,
@@ -58,40 +60,40 @@ def mttrpo_metaworld_mt10(ctxt, seed, entropy):
 
         baseline = GaussianMLPBaseline(
             env_spec=env.spec,
-            hidden_sizes=(256, 256),
+            hidden_sizes=(512, 512),
             use_trust_region=True,
         )
 
         sampler = RaySampler(agents=policy,
                              envs=envs,
                              max_episode_length=env.spec.max_episode_length,
-                             n_workers=10,
+                             n_workers=50,
                              is_tf_worker=True)
 
-        algo = TRPO(
-            env_spec=env.spec,
-            policy=policy,
-            baseline=baseline,
-            discount=0.99,
-            gae_lambda=0.95,
-            lr_clip_range=0.2,
-            stop_entropy_gradient=True,
-            entropy_method='max',
-            policy_ent_coeff=entropy,
-            center_adv=False,
-            use_softplus_entropy=False,
-            sampler=sampler,
-            use_neg_logli_entropy=True,
-            multitask=True,
-            train_task_sampler=train_task_sampler,
-            num_tasks=10,
-            task_update_frequency=50,
-        )
+        algo = PPO(env_spec=env.spec,
+                   policy=policy,
+                   baseline=baseline,
+                   discount=0.99,
+                   gae_lambda=0.95,
+                   lr_clip_range=0.2,
+                   optimizer=FirstOrderOptimizer,
+                   optimizer_args=dict(
+                       learning_rate=5e-4,
+                       max_optimization_epochs=16,
+                   ),
+                   stop_entropy_gradient=True,
+                   entropy_method='max',
+                   policy_ent_coeff=entropy,
+                   center_adv=False,
+                   use_softplus_entropy=False,
+                   sampler=sampler,
+                   use_neg_logli_entropy=True,
+                   multitask=True)
 
-        trainer.setup(algo, env)
-        trainer.train(n_epochs=int(500000000 / (500 * 100)),
+        trainer.setup(algo, envs)
+        trainer.train(n_epochs=int(100000000 / (500 * 100)),
                       batch_size=2,
                       plot=False)
 
 
-mttrpo_metaworld_mt10()
+mtppo_metaworld_mt1()
