+++ metaworld_examples/mtsac_metaworld_mt1.py
@@ -10,12 +10,10 @@ from torch import nn
 from torch.nn import functional as F
 
 from garage import wrap_experiment
-from garage.envs import normalize
 from garage.experiment import deterministic
 from garage.experiment.task_sampler import MetaWorldTaskSampler
 from garage.replay_buffer import PathBuffer
-from garage.sampler import FragmentWorker, LocalSampler
-from garage.sampler.default_worker import DefaultWorker
+from garage.sampler import DefaultWorker, EvalWorker, RaySampler
 from garage.torch import set_gpu_mode
 from garage.torch.algos import MTSAC
 from garage.torch.policies import TanhGaussianMLPPolicy
@@ -25,7 +23,7 @@ from garage.trainer import Trainer
 
 @click.command()
 @click.option('--seed', 'seed', type=int, default=1)
-@click.option('--env-name')
+@click.option('--env-name', type=str, default='pick-place-v2')
 @wrap_experiment(snapshot_mode='none', name_parameters='passed')
 def mtsac_metaworld_mt1(ctxt=None, *, seed, env_name):
     """Train MTSAC with mt1 environment.
@@ -35,9 +33,8 @@ def mtsac_metaworld_mt1(ctxt=None, *, seed, env_name):
             configuration used by Trainer to create the snapshotter.
         seed (int): Used to seed the random number generator to produce
             determinism.
-        _gpu (int): The ID of the gpu to be used (used on multi-gpu machines).
-        num_tasks (int): Number of tasks to use. Should be a multiple of 10.
-        timesteps (int): Number of timesteps to run.
+        env_name (str): Name of Meta-World environment to initialize
+            this experiment with.
 
     """
     _gpu = 0
@@ -53,11 +50,9 @@ def mtsac_metaworld_mt1(ctxt=None, *, seed, env_name):
                                               add_env_onehot=True)
 
     assert num_tasks % 50 == 0
-    assert num_tasks <= 500
+    assert num_tasks <= 50
     mt1_train_envs = train_task_sampler.sample(num_tasks)
     env = mt1_train_envs[0]()
-    mt1_test_envs = [env_up() for env_up in mt1_train_envs]
-
     policy = TanhGaussianMLPPolicy(
         env_spec=env.spec,
         hidden_sizes=[400, 400],
@@ -76,42 +71,46 @@ def mtsac_metaworld_mt1(ctxt=None, *, seed, env_name):
                                  hidden_nonlinearity=F.relu)
 
     replay_buffer = PathBuffer(capacity_in_transitions=int(1e6), )
+    max_episode_length = env.spec.max_episode_length
 
-    sampler = LocalSampler(
+    sampler = RaySampler(
         agents=policy,
         envs=mt1_train_envs,
-        max_episode_length=env.spec.max_episode_length,
+        max_episode_length=max_episode_length,
         # 1 sampler worker for each environment
         n_workers=num_tasks,
-        worker_class=FragmentWorker,
-        # increasing n_envs increases the vectorization of a sampler worker
-        # which improves runtime performance, but you will need to adjust this
-        # depending on your memory constraints. For reference, each worker by
-        # default uses n_envs=8. Each environment is approximately ~50mb large
-        # so creating 50 envs with 8 copies comes out to 20gb of memory. Many
-        # users want to be able to run multiple seeds on 1 machine, so I have
-        # reduced this to n_envs = 2 for 2 copies in the meantime.
-        worker_args=dict(n_envs=2))
+        worker_class=DefaultWorker)
+
+    test_sampler = RaySampler(
+        agents=policy,
+        envs=mt1_train_envs,
+        max_episode_length=max_episode_length,
+        # 1 sampler worker for each environment
+        n_workers=num_tasks,
+        worker_class=EvalWorker)
 
     batch_size = int(env.spec.max_episode_length * num_tasks)
     num_evaluation_points = 500
     epochs = timesteps // batch_size
     epoch_cycles = epochs // num_evaluation_points
     epochs = epochs // epoch_cycles
-    mtsac = MTSAC(policy=policy,
-                  qf1=qf1,
-                  qf2=qf2,
-                  sampler=sampler,
-                  gradient_steps_per_itr=env.spec.max_episode_length,
-                  eval_env=mt1_test_envs,
-                  env_spec=env.spec,
-                  num_tasks=num_tasks,
-                  steps_per_epoch=1,
-                  replay_buffer=replay_buffer,
-                  min_buffer_size=25000,
-                  target_update_tau=5e-3,
-                  discount=0.99,
-                  buffer_batch_size=7500)
+    mtsac = MTSAC(
+        policy=policy,
+        qf1=qf1,
+        qf2=qf2,
+        sampler=sampler,
+        test_sampler=test_sampler,
+        gradient_steps_per_itr=max_episode_length,
+        train_task_sampler=train_task_sampler,
+        env_spec=env.spec,
+        num_tasks=num_tasks,
+        #   steps_per_epoch=epoch_cycles,
+        steps_per_epoch=1,
+        replay_buffer=replay_buffer,
+        min_buffer_size=max_episode_length * num_tasks,
+        target_update_tau=5e-3,
+        discount=0.99,
+        buffer_batch_size=7500)
     if _gpu is not None:
         set_gpu_mode(True, _gpu)
     mtsac.to()
