+++ metaworld_examples/ppo_metaworld.py
@@ -6,19 +6,15 @@ steps.
 
 """
 
-import pickle
-
 import click
-import gym
-from metaworld.envs.mujoco.env_dict import (ALL_V1_ENVIRONMENTS,
-                                            ALL_V2_ENVIRONMENTS)
+from metaworld.envs import ALL_V2_ENVIRONMENTS_GOAL_OBSERVABLE
 import numpy as np
 import tensorflow as tf
 
 from garage import wrap_experiment
 from garage.envs import GymEnv, normalize
 from garage.experiment.deterministic import set_seed
-from garage.sampler import LocalSampler, RaySampler
+from garage.sampler import RaySampler
 from garage.tf.algos import PPO
 from garage.tf.baselines import GaussianMLPBaseline
 from garage.tf.optimizers import FirstOrderOptimizer
@@ -30,48 +26,29 @@ from garage.trainer import TFTrainer
 @click.option('--env-name', type=str, default='push-v2')
 @click.option('--seed', type=int, default=np.random.randint(0, 1000))
 @click.option('--entropy', type=float, default=0.01)
-@click.option('--use_softplus_entropy', type=bool, default=False)
-@click.option('--extra_tags', type=str, default='none')
 @wrap_experiment(name_parameters='all', snapshot_mode='gap', snapshot_gap=25)
-def ppo_metaworld(
-    ctxt=None,
-    env_name=None,
-    tag='',
-    extra_tags='',
-    entropy=5e-3,
-    stop_entropy_gradient=True,
-    use_softplus_entropy=False,
-    seed=1,
-):
+def ppo_metaworld(ctxt=None, env_name=None, entropy=5e-3, seed=1):
     """Train PPO with Metaworld environments.
 
     Args:
         ctxt (garage.experiment.ExperimentContext): The experiment
             configuration used by LocalRunner to create the snapshotter.
+        env_name (str): Name of Meta-World environment to initialize
+            this experiment with.
+        entropy (float): Coefficient to weigh the entropy reward term by
+            when using the max entropy reward.
         seed (int): Used to seed the random number generator to produce
             determinism.
 
     """
+    stop_entropy_gradient = True
+    use_softplus_entropy = False
     set_seed(seed)
-    special = {
-        'push-v1': 'push',
-        'reach-v1': 'reach',
-        'pick-place-v1': 'pick_place'
-    }
     not_in_mw = 'the env_name specified is not a metaworld environment'
-    assert env_name in ALL_V2_ENVIRONMENTS or env_name in ALL_V1_ENVIRONMENTS, not_in_mw
-
-    if env_name in ALL_V2_ENVIRONMENTS:
-        env_cls = ALL_V2_ENVIRONMENTS[env_name]
-    else:
-        env_cls = ALL_V1_ENVIRONMENTS[env_name]
-
-    env = env_cls()
-    env._partially_observable = False
-    env._freeze_rand_vec = False
-    env._set_task_called = True
-    env.reset()
-    env._freeze_rand_vec = True
+    env_name = env_name + '-goal-observable'
+    assert env_name in ALL_V2_ENVIRONMENTS_GOAL_OBSERVABLE, not_in_mw
+    env_cls = ALL_V2_ENVIRONMENTS_GOAL_OBSERVABLE[env_name]
+    env = env_cls(seed=seed)
     max_path_length = env.max_path_length
     env = normalize(GymEnv(env, max_episode_length=max_path_length),
                     normalize_reward=True)
@@ -96,11 +73,6 @@ def ppo_metaworld(
                              envs=env,
                              max_episode_length=env.spec.max_episode_length,
                              is_tf_worker=True)
-        # sampler = LocalSampler(agents=policy,
-        #                        envs=env,
-        #                        max_episode_length=env.spec.max_episode_length,
-        #                        is_tf_worker=True)
-
         # NOTE: make sure when setting entropy_method to 'max', set
         # center_adv to False and turn off policy gradient. See
         # tf.algos.NPO for detailed documentation.
